Chat_app/requirements.txt
Flask==3.1.2
flask-socketio==5.5.1
eventlet==0.40.3
gunicorn==23.0.0
python-engineio==4.12.3
python-socketio==5.14.3
requests==2.32.5
supabase==1.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
